getting rid of classnames
